brought back the old error handling
